move Collector class to _common
