+++ python/ray/_common/test_utils.py
@@ -178,3 +178,14 @@ def simulate_s3_bucket(
     yield url
     server.stop()
     os.environ = old_env
+
+@ray.remote
+class Collector:
+    def __init__(self):
+        self.items = []
+
+    def add(self, item):
+        self.items.append(item)
+
+    def get(self):
+        return self.items
