+++ python/ray/serve/tests/test_http_cancellation.py
@@ -8,8 +8,7 @@ from starlette.requests import Request
 
 import ray
 from ray import serve
-from ray._common.test_utils import SignalActor, wait_for_condition
-from ray._private.test_utils import Collector
+from ray._common.test_utils import SignalActor, Collector, wait_for_condition
 from ray.serve._private.test_utils import (
     get_application_url,
     send_signal_on_cancellation,
